# fe/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VAI Code",
    "description": "Weather Bridge AI production-oriented API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_v1_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "type": "array",
                  "title": "Response List Alerts Api V1 Alerts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Create Alert",
        "operationId": "create_alert_api_v1_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/submit": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Submit Alert",
        "operationId": "submit_alert_api_v1_alerts__alert_id__submit_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/publish": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Publish Alert",
        "operationId": "publish_alert_api_v1_alerts__alert_id__publish_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishAlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alerts/{alert_id}/acknowledgements": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Acknowledge Alert",
        "operationId": "acknowledge_alert_api_v1_alerts__alert_id__acknowledgements_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeAlertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertInboxItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/web-push/subscriptions": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Subscribe Web Push",
        "operationId": "subscribe_web_push_api_v1_notifications_web_push_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebPushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebPushSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/web-push/subscriptions/{contact_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Unsubscribe Web Push",
        "operationId": "unsubscribe_web_push_api_v1_notifications_web_push_subscriptions__contact_id__delete",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcknowledgeAlertRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "seen",
              "safe",
              "need_help"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AcknowledgeAlertRequest"
      },
...
      "AlertCreateRequest": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "evacuation"
            ],
            "title": "Source",
            "default": "manual"
          },
          "hazard_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "flash_flood",
                  "landslide",
                  "fog"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazard Type"
          },
          "level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Level"
          },
          "tier": {
            "type": "string",
            "enum": [
              "prepare",
              "go_now"
            ],
            "title": "Tier"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "what_happened": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "What Happened"
          },
          "danger_description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Danger Description"
          },
          "action_instruction": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Action Instruction"
          },
          "deadline_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "target_area_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Target Area Codes"
          }
        },
        "type": "object",
        "required": [
          "level",
          "tier",
          "confidence",
          "what_happened",
          "danger_description",
          "action_instruction",
          "deadline_at",
          "expires_at",
          "target_area_codes"
        ],
        "title": "AlertCreateRequest"
      },
      "AlertInboxItem": {
        "properties": {
          "alert_id": {
            "type": "string",
            "format": "uuid",
            "title": "Alert Id"
          },
          "recipient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recipient Id"
          },
          "hazard_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazard Type"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "what_happened": {
            "type": "string",
            "title": "What Happened"
          },
          "danger_description": {
            "type": "string",
            "title": "Danger Description"
          },
          "action_instruction": {
            "type": "string",
            "title": "Action Instruction"
          },
          "deadline_instruction": {
            "type": "string",
            "title": "Deadline Instruction"
          },
          "deadline_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At"
          },
          "acknowledgement_status": {
            "type": "string",
            "title": "Acknowledgement Status"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "recipient_id",
          "hazard_type",
          "level",
          "tier",
          "what_happened",
          "danger_description",
          "action_instruction",
          "deadline_instruction",
          "deadline_at",
          "acknowledgement_status",
          "acknowledged_at"
        ],
        "title": "AlertInboxItem"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "hazard_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazard Type"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "what_happened": {
            "type": "string",
            "title": "What Happened"
          },
          "danger_description": {
            "type": "string",
            "title": "Danger Description"
          },
          "action_instruction": {
            "type": "string",
            "title": "Action Instruction"
          },
          "deadline_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "target_area_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Area Codes"
          },
          "recipient_count": {
            "type": "integer",
            "title": "Recipient Count"
          },
          "delivery_count": {
            "type": "integer",
            "title": "Delivery Count"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "status",
          "hazard_type",
          "level",
          "tier",
          "confidence",
          "what_happened",
          "danger_description",
          "action_instruction",
          "deadline_at",
          "expires_at",
          "target_area_codes",
          "recipient_count",
          "delivery_count",
          "published_at"
        ],
        "title": "AlertResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishAlertResponse": {
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/AlertResponse"
          },
          "recipient_count": {
            "type": "integer",
            "title": "Recipient Count"
          },
          "delivery_count": {
            "type": "integer",
            "title": "Delivery Count"
          }
        },
        "type": "object",
        "required": [
          "alert",
          "recipient_count",
          "delivery_count"
        ],
        "title": "PublishAlertResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebPushKeys": {
        "properties": {
          "p256dh": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "WebPushKeys"
      },
      "WebPushSubscriptionRequest": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "expirationTime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expirationtime"
          },
          "keys": {
            "$ref": "#/components/schemas/WebPushKeys"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "WebPushSubscriptionRequest"
      },
      "WebPushSubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "is_active",
          "last_seen_at"
        ],
        "title": "WebPushSubscriptionResponse"
      }
    }
  }
}